[serve] Fold the actor-class factories into proxy_state / controller
get_proxy_actor_class and get_controller_impl were extracted into dedicated
modules (proxy_actor_class.py, controller_actor_class.py) to break the
default_impl import cycle. But RayTurbo does not actually override either (its
copies are byte-identical to OSS; the real Anyscale overrides are
AnyscaleClusterNodeInfoCache / AnyscaleDeploymentScheduler, whose leaf factories
stay in default_impl). So the dedicated extension-point modules aren't warranted.

Move the factories next to the actors they build instead:
- get_proxy_actor_class -> proxy_state.py (already imports ProxyActor; only
  imported by controller, so no cycle)
- get_controller_impl -> controller.py (next to ServeController)

Both still leave default_impl (required to import their actors eagerly), so the
cycle stays broken; this just drops two files. They can't share one module:
get_proxy_actor_class is imported by controller while get_controller_impl
imports controller.
